+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-* [855886f] Treat figure the same as img
+* [3a835b3] Disable Cypress/Composer/WordPress tests
+* [1892a6d] Use latest versions of upload-artifact and download-artifact actions
+* [1a5dce8] Hotfix for missing $variables caused by changes to frontend.scss in this commit
+* [3e73b50] Link Proxy - prevent feedback loop in iOS Safari